Quad value pick
The thing was before i was taking value from the sum and product of the quad eq from the user. but then i decide to take the value of a, b, c. so i used regular reaction library to rctract the value of a, b, c and convert them into sum and product or equartion